ml/gpt-vision-analysis/requirements.txt
openai>=1.12.0
pydantic>=2.6.1
python-dotenv>=1.0.0
pillow>=10.2.0
pytest>=8.0.0
httpx>=0.26.0
rich>=13.7.0 